@@ -5,9 +5,9 @@ build-backend = "pdm.backend"
 [project]
 authors = []
 license = { text = "MIT" }
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dependencies = [
-    "langchain-core<1.0.0,>=0.3.55",
+    "langchain-core<2.0.0,>=0.3.55",
     "fireworks-ai>=0.13.0",
     "openai<2.0.0,>=1.10.0",
     "requests<3,>=2",
