@@ -5,8 +5,8 @@ build-backend = "pdm.backend"
 [project]
 authors = []
 license = { text = "MIT" }
-requires-python = ">=3.9"
-dependencies = ["langchain-core<1.0.0,>=0.3.72", "groq<1,>=0.30.0"]
+requires-python = ">=3.10"
+dependencies = ["langchain-core<2.0.0,>=0.3.72", "groq<1,>=0.30.0"]
 name = "langchain-groq"
 version = "0.3.7"
 description = "An integration package connecting Groq and LangChain"
