@@ -5,9 +5,9 @@ build-backend = "pdm.backend"
 [project]
 authors = []
 license = { text = "MIT" }
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dependencies = [
-    "langchain-core<1.0.0,>=0.3.68",
+    "langchain-core<2.0.0,>=0.3.68",
     "tokenizers<1,>=0.15.1",
     "httpx<1,>=0.25.2",
     "httpx-sse<1,>=0.3.1",
