@@ -5,14 +5,14 @@ build-backend = "pdm.backend"
 [project]
 authors = []
 license = { text = "MIT" }
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dependencies = [
-    "langchain-core<1.0.0,>=0.3.74",
+    "langchain-core<2.0.0,>=1.0.0a1",
     "openai<2.0.0,>=1.99.9",
     "tiktoken<1,>=0.7",
 ]
 name = "langchain-openai"
-version = "0.3.32"
+version = "1.0.0a1"
 description = "An integration package connecting OpenAI and LangChain"
 readme = "README.md"
 
