@@ -6,7 +6,7 @@ import re
 import shutil
 from collections.abc import Sequence
 from pathlib import Path
-from typing import Optional, TypedDict
+from typing import Any, TypedDict
 
 from git import Repo
 
@@ -23,18 +23,18 @@ class DependencySource(TypedDict):
     """Dependency source information."""
 
     git: str
-    ref: Optional[str]
-    subdirectory: Optional[str]
-    api_path: Optional[str]
-    event_metadata: dict
+    ref: str | None
+    subdirectory: str | None
+    api_path: str | None
+    event_metadata: dict[str, Any]
 
 
 # use poetry dependency string format
 def parse_dependency_string(
-    dep: Optional[str],
-    repo: Optional[str],
-    branch: Optional[str],
-    api_path: Optional[str],
+    dep: str | None,
+    repo: str | None,
+    branch: str | None,
+    api_path: str | None,
 ) -> DependencySource:
     """Parse a dependency string into a DependencySource."""
     if dep is not None and dep.startswith("git+"):
@@ -112,7 +112,7 @@ def parse_dependency_string(
     )
 
 
-def _list_arg_to_length(arg: Optional[list[str]], num: int) -> Sequence[Optional[str]]:
+def _list_arg_to_length(arg: list[str] | None, num: int) -> Sequence[str | None]:
     if not arg:
         return [None] * num
     if len(arg) == 1:
@@ -124,7 +124,7 @@ def _list_arg_to_length(arg: Optional[list[str]], num: int) -> Sequence[Optional
 
 
 def parse_dependencies(
-    dependencies: Optional[list[str]],
+    dependencies: list[str] | None,
     repo: list[str],
     branch: list[str],
     api_path: list[str],
@@ -158,11 +158,12 @@ def parse_dependencies(
             inner_repos,
             inner_branches,
             inner_api_paths,
+            strict=False,
         )
     ]
 
 
-def _get_repo_path(gitstring: str, ref: Optional[str], repo_dir: Path) -> Path:
+def _get_repo_path(gitstring: str, ref: str | None, repo_dir: Path) -> Path:
     # only based on git for now
     ref_str = ref if ref is not None else ""
     hashed = hashlib.sha256((f"{gitstring}:{ref_str}").encode()).hexdigest()[:8]
@@ -176,7 +177,7 @@ def _get_repo_path(gitstring: str, ref: Optional[str], repo_dir: Path) -> Path:
     return repo_dir / directory_name
 
 
-def update_repo(gitstring: str, ref: Optional[str], repo_dir: Path) -> Path:
+def update_repo(gitstring: str, ref: str | None, repo_dir: Path) -> Path:
     """Update a git repository to the specified ref."""
     # see if path already saved
     repo_path = _get_repo_path(gitstring, ref, repo_dir)
