@@ -3,7 +3,6 @@
 import gzip
 from os import PathLike
 from pathlib import Path
-from typing import Union
 
 import pytest
 import yaml
@@ -48,7 +47,9 @@ class CustomPersister:
 
     @classmethod
     def load_cassette(
-        cls, cassette_path: Union[str, PathLike[str]], serializer: CustomSerializer
+        cls,
+        cassette_path: str | PathLike[str],
+        serializer: CustomSerializer,
     ) -> tuple[dict, dict]:
         """Load a cassette from a file."""
         # If cassette path is already Path this is a no-op
@@ -63,7 +64,7 @@ class CustomPersister:
 
     @staticmethod
     def save_cassette(
-        cassette_path: Union[str, PathLike[str]],
+        cassette_path: str | PathLike[str],
         cassette_dict: dict,
         serializer: CustomSerializer,
     ) -> None:
