@@ -1,7 +1,5 @@
 """Test the standard tests on the custom chat model in the docs."""
 
-from typing import Optional
-
 import pytest
 from langchain_core.language_models.chat_models import BaseChatModel
 
@@ -34,7 +32,7 @@ class TestChatParrotLinkIntegration(ChatModelIntegrationTests):
     def test_unicode_tool_call_integration(
         self,
         model: BaseChatModel,
-        tool_choice: Optional[str] = None,
+        tool_choice: str | None = None,
         force_tool_call: bool = True,
     ) -> None:
         """Expected failure as ChatParrotLink doesn't support tool calling yet."""
