@@ -33,9 +33,11 @@ from pydantic import Discriminator, Field, Tag
 from langchain_core.exceptions import ErrorCode, create_message
 from langchain_core.messages.ai import AIMessage, AIMessageChunk
 from langchain_core.messages.base import BaseMessage, BaseMessageChunk
+from langchain_core.messages.block_translators.openai import (
+    convert_to_openai_data_block,
+)
 from langchain_core.messages.chat import ChatMessage, ChatMessageChunk
 from langchain_core.messages.content import (
-    convert_to_openai_data_block,
     is_data_content_block,
 )
 from langchain_core.messages.function import FunctionMessage, FunctionMessageChunk
